New detour (0xa0fa40/0xa0fa30)
This function is called inside the main original detour function (0xa3ba200/0xa3b350), and it is in part used by said function to determine which sort of creature it is dealing with. Basically, we are fooling the main function to think it's dealing with a tribal creature, even when the creature in question is civilian or above. (Probably need to think if same precautions as previously are going to be needed here)